:recycle: add 4 decimals precision for liquidation incentives

+++ test/utils/Events.sol
@@ -52,8 +52,7 @@ abstract contract Events {
                             LIQUIDATOR
     ////////////////////////////////////////////////////////////// */
 
-    // note: to delete when removing old liquidator contract
-    event WeightsSet(uint8 initiatorRewardWeight, uint8 penaltyWeight, uint8 closingRewardWeight);
+    event WeightsSet(uint16 initiatorRewardWeight, uint16 penaltyWeight, uint16 closingRewardWeight);
     event AuctionCurveParametersSet(uint64 base, uint32 cutoffTime);
     event StartPriceMultiplierSet(uint16 startPriceMultiplier);
     event MinimumPriceMultiplierSet(uint8 minPriceMultiplier);
